# message-mesh-svc/src/main/resources/message-mesh-openapi.yaml
{
  "openapi": "3.0.1",
  "info": {
    "title": "Message Mesh API",
...
    "contact": {
      "name": "Message Mesh Platform Team",
      "url": "https://message-mesh.dev.com",
      "email": "platform@message-mesh.dev.com"
    },
    "license": {
      "name": "Apache License 2.0",
      "url": "https://www.apache.org/licenses/LICENSE-2.0"
    },
    "version": "1.0.0"
  },
  "servers": [
    {
      "url": "http://localhost:8080",
      "description": "Local development server"
    },
    {
      "url": "https://api.message-mesh.prod.com",
      "description": "Production server"
    }
  ],
  "security": [
    {
      "bearerAuth": []
    }
  ],
...
  "paths": {
...
    "/api/auth/register": {
      "post": {
        "tags": [
          "Authentication"
        ],
        "summary": "Register a new account",
        "description": "Creates a user with a unique username and returns a signed JWT together with the created profile. The returned token authorizes all subsequent REST calls and the WebSocket handshake. No authentication is required to call this endpoint.",
        "operationId": "register",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RegisterRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "409": {
            "description": "Username already taken",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "201": {
            "description": "Account created; JWT issued",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AuthResponse"
                }
              }
            }
          },
          "400": {
            "description": "Validation failed (e.g. missing or too-short fields)",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          }
        },
        "security": []
      }
    },
    "/api/auth/login": {
      "post": {
        "tags": [
          "Authentication"
        ],
        "summary": "Authenticate and obtain a JWT",
        "description": "Validates the supplied credentials and, on success, returns a signed JWT and the user profile. Use the token as 'Authorization: Bearer \u003Ctoken\u003E' for protected endpoints.",
        "operationId": "login",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/LoginRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "401": {
            "description": "Invalid username or password",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "400": {
            "description": "Missing username or password",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "200": {
            "description": "Credentials accepted; JWT issued",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AuthResponse"
                }
              }
            }
          }
        },
        "security": []
      }
    },
...
    "/api/conversations/{id}/messages": {
      "get": {
        "tags": [
          "Conversations"
        ],
        "summary": "Fetch message history",
        "description": "Returns messages for a conversation ordered by ascending sequence number. Use 'afterSeq' for forward, cursor-style pagination: pass the highest 'seq' you already hold to fetch only newer messages, and bound the page with 'limit'.",
        "operationId": "messages",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "Conversation identifier",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          },
          {
            "name": "afterSeq",
            "in": "query",
            "description": "Return only messages with a sequence greater than this value (cursor)",
            "required": false,
            "schema": {
              "type": "integer",
              "format": "int64",
              "default": 0
            },
            "example": 0
          },
          {
            "name": "limit",
            "in": "query",
            "description": "Maximum number of messages to return",
            "required": false,
            "schema": {
              "type": "integer",
              "format": "int32",
              "default": 50
            },
            "example": 50
          }
        ],
        "responses": {
          "403": {
            "description": "Caller is not a member of the conversation",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "200": {
            "description": "Messages returned",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/MessageDto"
                  }
                }
              }
            }
          },
          "401": {
            "description": "Missing or invalid JWT",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "404": {
            "description": "Conversation not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/admin/stats": {
      "get": {
        "tags": [
          "Admin"
        ],
        "summary": "Get platform statistics",
        "description": "Returns aggregate counts for users, conversations, messages and audit activity used to populate the admin dashboard.",
        "operationId": "stats",
        "responses": {
          "200": {
            "description": "Statistics returned",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AdminStatsDto"
                }
              }
            }
          },
          "403": {
            "description": "Caller is not an administrator",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/admin/conversations/{id}/messages": {
      "get": {
        "tags": [
          "Admin"
        ],
        "summary": "List messages in a conversation",
        "description": "Returns a page of messages (newest first) for moderation review, including soft-deleted tombstones. 'page' (0-based) and 'size' (max 100) paginate.",
        "operationId": "messages_1",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          },
          {
            "name": "page",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "format": "int32",
              "default": 0
            }
          },
          {
            "name": "size",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "format": "int32",
              "default": 50
            }
          }
        ],
        "responses": {
          "403": {
            "description": "Caller is not an administrator",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "200": {
            "description": "Messages returned",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PagedResponseMessageDto"
                }
              }
            }
          },
          "404": {
            "description": "Conversation not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "ApiError": {
        "type": "object",
        "properties": {
          "timestamp": {
            "type": "string",
            "format": "date-time"
          },
          "status": {
            "type": "integer",
            "format": "int32"
          },
          "error": {
            "type": "string"
          },
          "message": {
            "type": "string"
          },
          "requestId": {
            "type": "string"
          },
          "fieldErrors": {
            "type": "object",
            "additionalProperties": {
              "type": "string"
            }
          }
        }
      },
      "AuthResponse": {
        "type": "object",
        "properties": {
          "token": {
            "type": "string"
          },
          "user": {
            "$ref": "#/components/schemas/UserDto"
          }
        }
      },
      "UserDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "username": {
            "type": "string"
          },
          "displayName": {
            "type": "string"
          },
          "role": {
            "type": "string",
            "enum": [
              "USER",
              "ADMIN"
            ]
          }
        }
      },
...
      "MessageDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "conversationId": {
            "type": "string",
            "format": "uuid"
          },
          "senderUsername": {
            "type": "string"
          },
          "seq": {
            "type": "integer",
            "format": "int64"
          },
          "body": {
            "type": "string"
          },
          "status": {
            "type": "string",
            "enum": [
              "SENT",
              "DELIVERED",
              "READ"
            ]
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          },
          "clientTempId": {
            "type": "string"
          },
          "editedAt": {
            "type": "string",
            "format": "date-time"
          },
          "deleted": {
            "type": "boolean"
          },
          "parentId": {
            "type": "string",
            "format": "uuid"
          },
          "parentPreview": {
            "$ref": "#/components/schemas/ParentPreview"
          },
          "reactions": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ReactionSummary"
            }
          }
        }
      },
      "ParentPreview": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "senderUsername": {
            "type": "string"
          },
          "body": {
            "type": "string"
          }
        }
      },
      "ReactionSummary": {
        "type": "object",
        "properties": {
          "emoji": {
            "type": "string"
          },
          "count": {
            "type": "integer",
            "format": "int32"
          },
          "usernames": {
            "type": "array",
            "items": {
              "type": "string"
            }
          }
        }
      },
...
      "RegisterRequest": {
        "required": [
          "displayName",
          "password",
          "username"
        ],
        "type": "object",
        "properties": {
          "username": {
            "maxLength": 64,
            "minLength": 3,
            "type": "string"
          },
          "password": {
            "maxLength": 100,
            "minLength": 6,
            "type": "string"
          },
          "displayName": {
            "maxLength": 128,
            "minLength": 0,
            "type": "string"
          }
        }
      },
      "LoginRequest": {
        "required": [
          "password",
          "username"
        ],
        "type": "object",
        "properties": {
          "username": {
            "type": "string"
          },
          "password": {
            "type": "string"
          }
        }
      },
...
      "PagedResponseMessageDto": {
        "type": "object",
        "properties": {
          "content": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/MessageDto"
            }
          },
          "page": {
            "type": "integer",
            "format": "int32"
          },
          "size": {
            "type": "integer",
            "format": "int32"
          },
          "totalElements": {
            "type": "integer",
            "format": "int64"
          },
          "totalPages": {
            "type": "integer",
            "format": "int32"
          }
        }
      },
...
      "AdminStatsDto": {
        "type": "object",
        "properties": {
          "totalUsers": {
            "type": "integer",
            "format": "int64"
          },
          "activeUsers": {
            "type": "integer",
            "format": "int64"
          },
          "inactiveUsers": {
            "type": "integer",
            "format": "int64"
          },
          "adminUsers": {
            "type": "integer",
            "format": "int64"
          },
          "onlineUsers": {
            "type": "integer",
            "format": "int64"
          },
          "newUsersLast7Days": {
            "type": "integer",
            "format": "int64"
          },
          "totalConversations": {
            "type": "integer",
            "format": "int64"
          },
          "groupConversations": {
            "type": "integer",
            "format": "int64"
          },
          "directConversations": {
            "type": "integer",
            "format": "int64"
          },
          "deletedConversations": {
            "type": "integer",
            "format": "int64"
          },
          "totalMessages": {
            "type": "integer",
            "format": "int64"
          },
          "messagesLast7Days": {
            "type": "integer",
            "format": "int64"
          },
          "auditEvents": {
            "type": "integer",
            "format": "int64"
          }
        }
      },
...
    },
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "description": "JWT bearer token obtained from the authentication endpoints. Paste only the raw token value; Swagger UI adds the 'Bearer ' prefix.",
        "name": "bearerAuth",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    }
  }
}